+++ app/league/api/views.py
@@ -5,9 +5,9 @@ from datetime import timezone
 from flask import Blueprint, jsonify, request, url_for
 from flask_login import login_required
 
-from league.models import Color, Game, Player
 from league.api.forms import GameCreateForm, GameUpdateForm
 from league.extensions import csrf_protect, messenger
+from league.models import Color, Game, Player
 
 blueprint = Blueprint('api', __name__, url_prefix='/api/v1.0',
                       static_folder='../static')
